// backend/validation-dtos/auth.dto.ts
import { z } from "zod";

const signupDto = z.object({
    username: z.string({ required_error: "Username is required" }),
    fullname: z.string({ required_error: "Fullname is required" }),
    email: z.string({ required_error: "Email is required" }).email({ message: "Invalid email" }),
    password: z.string({ required_error: "Password is required" }).min(6, { message: "Password must be at least 6 characters" }),
});


const signinDto = z.object({
    username: z.string({ required_error: "Username is required" }),
    password: z.string({ required_error: "Password is required" }).min(6, { message: "Password must be at least 6 characters" }),
})


export { signupDto, signinDto }